Made fully work app

+++ services/courier/pkg/migrations/0001_create_courier_tables.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS couriers (
     id INT PRIMARY KEY,
     active BOOL NOT NULL,
     in_progress BOOL NOT NULL,
-    rating INT NOT NULL,
+    rating DECIMAL(10, 2) NOT NULL,
     order_delivered INT NOT NULL
 );
